requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1673535674859/work
aiohttp @ file:///C:/ci_311/aiohttp_1676432932774/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
altair==4.2.2
asttokens==2.2.1
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1681349778161/work
brotlipy==0.7.0
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1674482203741/work
catboost==1.2
certifi==2022.12.7
cffi @ file:///C:/ci_311/cffi_1676423759166/work
chardet==5.1.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///D:/bld/click_1666798499870/work
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
cryptography @ file:///C:/ci_311/cryptography_1679419210767/work
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
frozenlist @ file:///C:/ci_311/frozenlist_1676428131576/work
gitdb==4.0.10
GitPython==3.1.31
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1683751600794/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1675825735130/work
graphviz==0.20.1
grpcio @ file:///C:/b/abs_ccut8qek12/croot/grpc-suite_1681912611694/work
hyperlink==21.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
incremental==22.10.0
ipykernel==6.19.2
ipython==8.7.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
lxml==4.9.2
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1679584000376/work
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
multidict @ file:///C:/ci_311/multidict_1676428396308/work
nest-asyncio==1.5.6
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1666056362788/work
packaging==22.0
pandas==2.0.1
parsel==1.8.1
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.0
plotly==5.14.1
prompt-toolkit==3.0.36
Protego==0.2.1
protobuf @ file:///D:/bld/protobuf_1681050297321/work/python
psutil==5.9.4
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydeck==0.8.1b0
PyDispatcher==2.0.7
Pygments==2.13.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1683676063469/work
Pympler==1.0.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1680037383858/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.8.2
pytz==2023.3
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
pywin32==305
pyzmq==24.0.1
queuelib==1.6.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1682535435083/work
requests-file==1.5.1
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rich==13.3.5
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.8.0
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.22.0
tenacity==8.2.2
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1683288318191/work/tensorboard-2.13.0-py3-none-any.whl#sha256=ab69961ebddbddc83f5fa2ff9233572bdad5b883778c35e4fe94bf1798bd8481
tensorboard-data-server @ file:///D:/bld/tensorboard-data-server_1676292040279/work/tensorboard_data_server-0.7.0-py3-none-any.whl
threadpoolctl==3.1.0
tldextract==3.4.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.7.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
tzlocal==5.0.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
validators==0.20.0
w3lib==2.1.1
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
yarl @ file:///C:/ci_311/yarl_1676432870023/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
zope.interface==6.0
